split large file test to not run with race detector, sine that's too big for CI
This commit was moved from ipfs/go-unixfsnode@05a4b1622c38f2e64eb40211e4ca19c0d524fa89
